setup.sh 1.4 (space in name/path supported)

+++ res/src/set.c
@@ -5,7 +5,7 @@ ud_arr      *ud_arr_set(size_t type_size, size_t len, ...)
     ud_arr      *new_arr;
     va_list     va;
     char        *set_val;
-    int         set_val_tmp;
+    long double set_val_tmp;
 
     UD_UT_PROT_MALLOC(new_arr = ud_arr_init(type_size, len));
     if (!len) return new_arr;
@@ -13,7 +13,7 @@ ud_arr      *ud_arr_set(size_t type_size, size_t len, ...)
     va_start(va, len);    
     while (len-- > 0)
     {
-        set_val_tmp = (int)va_arg(va, long long);
+        set_val_tmp = (long double)va_arg(va, long double);
         set_val = (char*)&set_val_tmp;
         for (ud_ut_count i = 0; i < type_size; ++i)
             *val++ = *set_val++;
